+++ helpers/escaper.js
@@ -1,0 +1,8 @@
+var Handlebars = require('handlebars');
+module.exports = function(text, options) {
+  text = text.string || text;
+
+  var newText = Handlebars.Utils.escapeExpression(text);
+
+  return new Handlebars.SafeString(newText);
+};